update import paths for reorganized domain files
- Update all import paths to match new directory structure
- Move test files to corresponding subdirectories
- Update relative paths in moved test files
- Ensure all imports point to correct locations:
  - discord-tokens/DiscordTokens.ts
  - user/User.ts and user/UserError.ts
  - appreciation/Appreciation.ts and appreciation/AppreciationError.ts
  - weekly-available-points/WeeklyAvailablePoints.ts

This completes the domain file reorganization with proper import path updates.
