+++ packages/backend/__tests__/testing/table_assert/AssertEqualUserTable.ts
@@ -1,6 +1,6 @@
 import { expect } from "vitest";
 import type * as schema from "../../../database/schema";
-import type { User } from "../../../src/domain/User";
+import type { User } from "../../../src/domain/user/User";
 
 /**
  * データベースのuserテーブルと引数で渡されたUserドメインオブジェクトが等しいことをアサート
