+++ packages/backend/src/application/usecases/auth.ts
@@ -1,7 +1,7 @@
 import type { Context } from "hono";
 import { inject, injectable } from "inversify";
-import { DiscordTokens } from "../../domain/DiscordTokens";
-import { DiscordID, User } from "../../domain/User";
+import { DiscordTokens } from "../../domain/discord-tokens/DiscordTokens";
+import { DiscordID, User } from "../../domain/user/User";
 import { TYPES } from "../../infrastructure/config/types";
 import type { DiscordTokensRepositoryInterface } from "../../infrastructure/repositories/DiscordTokensRepository";
 import type { UserRepositoryInterface } from "../../infrastructure/repositories/UserRepository";
